+++ src/AbstractIntegration.php
@@ -5,8 +5,9 @@
  * Description:
  * Copyright: Copyright (c) 2005 - 2016
  * Company: Pronamic
+ *
  * @author Remco Tolsma
- * @version 1.0.0
+ * @version 1.0.1
  * @since 1.0.0
  * @see https://github.com/thephpleague/omnipay-common/blob/master/src/Omnipay/Common/AbstractGateway.php
  */
@@ -29,5 +30,15 @@ abstract class Pronamic_WP_Pay_Gateways_AbstractIntegration implements Pronamic_
 
 	public function set_name( $name ) {
 		$this->name = $name;
+	}
+
+	/**
+	 * Get required settings for this integration.
+	 *
+	 * @see https://github.com/wp-premium/gravityforms/blob/1.9.16/includes/fields/class-gf-field-multiselect.php#L21-L42
+	 * @return array
+	 */
+	public function get_settings() {
+		return array();
 	}
 }