+++ src/IntegrationInterface.php
@@ -5,8 +5,9 @@
  * Description:
  * Copyright: Copyright (c) 2005 - 2016
  * Company: Pronamic
+ *
  * @author Remco Tolsma
- * @version 1.0.0
+ * @version 1.0.1
  * @since 1.0.0
  * @see https://github.com/thephpleague/omnipay-common/blob/master/src/Omnipay/Common/GatewayInterface.php
  */
@@ -17,9 +18,5 @@ interface Pronamic_WP_Pay_Gateways_IntegrationInterface {
 
 	public function get_config_factory_class();
 
-	public function get_config_class();
-
 	public function get_settings_class();
-
-	public function get_gateway_class();
 }