+++ CHANGELOG.md
@@ -7,7 +7,12 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 - 
 
+## [1.0.1] - 2016-03-02
+- Added get settings function.
+- Removed get_config_class() function from integration interfase.
+
 ## 1.0.0 - 2015-01-29
 - First release.
 
-[unreleased]: https://github.com/wp-pay-gateways/common/compare/1.0.0...HEAD
+[unreleased]: https://github.com/wp-pay-gateways/common/compare/1.0.1...HEAD
+[1.0.1]: https://github.com/wp-pay-gateways/common/compare/1.0.0...1.0.1